Added Profile
- Profile Picture
- Interest
- Gender
- Preferences
- Improved UI in chatroom

Need to add backend for profile page and reflect it in chatroom. Need to improve the logic for matching. There are some bugs especially when multiple users simultaneously queuing. I think we need something like a websocket (like a server function) idk, to handle the matching to only one function, not on each devices.